chore: adding firebase config files
