+++ firebase/config.ts
@@ -1,0 +1,19 @@
+import firebase from "firebase/app";
+import "firebase/auth";
+
+// Firebase config
+const firebaseConfig = {
+  apiKey: "",
+  authDomain: "",
+  projectId: "",
+  storageBucket: "",
+  messagingSenderId: "",
+  appId: ""
+};
+
+if (!firebase.apps.length) {
+  firebase.initializeApp(firebaseConfig);
+}
+
+export const auth = firebase.auth();
+export { firebase };
